requirements.txt
# Core LLM & Agent Framework
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langgraph>=0.2.0

# MCP Protocol
mcp[cli]>=1.0.0
langchain-mcp-adapters>=0.1.0

# Search Tools
duckduckgo-search>=6.0.0
tavily-python>=0.3.0

# UI & Visualization
streamlit>=1.40.0
plotly>=5.24.0
pandas>=2.0.0

# Utilities
python-dotenv>=1.0.0
rich>=13.0.0
pydantic>=2.0.0
httpx>=0.27.0
aiofiles>=23.0.0

# Database
aiosqlite>=0.20.0
